prepare counter app for modularization
