+++ src/apps/counter/counter.h
@@ -1,0 +1,10 @@
+#ifndef COUNTER_H
+#define COUNTER_H
+
+#include "buttons.h"
+
+void ButtonHandler(const Button *btn, ButtonEvent_t event);
+
+void RunApp(void);
+
+#endif // COUNTER_H
